pyproject.toml
[project]
name = "Coin-proto"
version = "0.4.0"
description = "This prototype is a cryptocurrency trading automation system."
authors = [
    {name = "lucas_hub12", email = "hyeonggyunbae@gmail.com"},
]
dependencies = ["typer>=0.15.1", "pyupbit>=0.2.34", "python-dotenv>=1.0.1", "streamlit>=1.42.0", "psycopg[binary,pool]>=3.2.4"]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[project.scripts]
coin-proto = "coin_proto.where_now:entry_point"

[dependency-groups]
test = [
    "pytest>=8.3.4",
]
note = [
    "jupyterlab>=4.3.4",
]
